requirements.txt
django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8